feat: added optional WithKeepAlive and WithRetryDelay (#107)
Some servers require the client to send ping to prevent the connection from being broken due to idling. We have added WithKeepAlive which accepts time.Duration will be used as an interval for doing a continuous ping to the server. This is also implemented in npm package graphql-ws.

WithRetryDelay makes retry delay customizable. If not set it will use the default (currently hard coded 1 second).

---------

Co-authored-by: applyinnovations <alexander@applyinnovations.com.au>